Moved substring functionality to new db_helper file
